Refactor AppShell and UserNav components to streamline imports and remove unused sidebar elements
